Refactoring
- Simplification pour l'ajout de solutions avec cmake.
- Les fichiers de la library common sont maintenant dans le bon répertoire.
- Nouveaux batch files pour la creation des rérpertoires de builds
- Changement de style pour avoir 120 colones

+++ src/common/md5.hpp
@@ -36,8 +36,8 @@ public:
 
         for (unsigned char b : hashBytes)
         {
-            hashText[offset++] = kDigits[b >> 4];
-            hashText[offset++] = kDigits[b & 0xf];
+            hashText[offset++] = kDigits[static_cast<std::size_t>(b >> 4)];
+            hashText[offset++] = kDigits[static_cast<std::size_t>(b & 0xf)];
         }
 
         return hashText;
